Move `TaskPoolPlugin` to `bevy_app`
