--- crates/bevy_core/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "bevy_core"
-version = "0.15.0-dev"
-edition = "2021"
-description = "Provides core functionality for Bevy Engine"
-homepage = "https://bevyengine.org"
-repository = "https://github.com/bevyengine/bevy"
-license = "MIT OR Apache-2.0"
-keywords = ["bevy"]
-
-[dependencies]
-# bevy
-bevy_app = { path = "../bevy_app", version = "0.15.0-dev", default-features = false }
-bevy_ecs = { path = "../bevy_ecs", version = "0.15.0-dev", default-features = false }
-bevy_reflect = { path = "../bevy_reflect", version = "0.15.0-dev", features = [
-  "bevy",
-], optional = true }
-bevy_tasks = { path = "../bevy_tasks", version = "0.15.0-dev" }
-bevy_utils = { path = "../bevy_utils", version = "0.15.0-dev" }
-
-# other
-serde = { version = "1.0", optional = true }
-
-[features]
-default = ["bevy_reflect"]
-bevy_reflect = [
-  "dep:bevy_reflect",
-  "bevy_app/bevy_reflect",
-  "bevy_ecs/bevy_reflect",
-]
-serialize = ["dep:serde"]
-
-[dev-dependencies]
-crossbeam-channel = "0.5.0"
-serde_test = "1.0"
-
-[lints]
-workspace = true
-
-[package.metadata.docs.rs]
-rustdoc-args = ["-Zunstable-options", "--generate-link-to-definition"]
-all-features = true
