--- crates/bevy_core/README.md
@@ -1,7 +1,0 @@
-# Bevy Core
-
-[![License](https://img.shields.io/badge/license-MIT%2FApache-blue.svg)](https://github.com/bevyengine/bevy#license)
-[![Crates.io](https://img.shields.io/crates/v/bevy_core.svg)](https://crates.io/crates/bevy_core)
-[![Downloads](https://img.shields.io/crates/d/bevy_core.svg)](https://crates.io/crates/bevy_core)
-[![Docs](https://docs.rs/bevy_core/badge.svg)](https://docs.rs/bevy_core/latest/bevy_core/)
-[![Discord](https://img.shields.io/discord/691052431525675048.svg?label=&logo=discord&logoColor=ffffff&color=7389D8&labelColor=6A7EC2)](https://discord.gg/bevy)
