+++ crates/bevy_internal/src/default_plugins.rs
@@ -5,7 +5,7 @@ plugin_group! {
     pub struct DefaultPlugins {
         bevy_app:::PanicHandlerPlugin,
         bevy_log:::LogPlugin,
-        bevy_core:::TaskPoolPlugin,
+        bevy_app:::TaskPoolPlugin,
         bevy_diagnostic:::FrameCountPlugin,
         bevy_time:::TimePlugin,
         bevy_transform:::TransformPlugin,
@@ -106,7 +106,7 @@ impl Plugin for IgnoreAmbiguitiesPlugin {
 plugin_group! {
     /// This plugin group will add the minimal plugins for a *Bevy* application:
     pub struct MinimalPlugins {
-        bevy_core:::TaskPoolPlugin,
+        bevy_app:::TaskPoolPlugin,
         bevy_diagnostic:::FrameCountPlugin,
         bevy_time:::TimePlugin,
         bevy_app:::ScheduleRunnerPlugin,
