+++ crates/bevy_tasks/src/usages.rs
@@ -92,7 +92,7 @@ taskpool! {
     (IO_TASK_POOL, IoTaskPool)
 }
 
-/// A function used by `bevy_core` to tick the global tasks pools on the main thread.
+/// A function used by `bevy_app` to tick the global tasks pools on the main thread.
 /// This will run a maximum of 100 local tasks per executor per call to this function.
 ///
 /// # Warning
